# applespi: Re-enable SPI on resume