@cb22 2016-09-12
Hi Cameron - thanks for the pull request! Quick question, what hardware / configuration have you been testing on? The reason I ask is that I've been struggling a bit with suspend / resume in general.

I'm on the MacBook9,1 and booting off the internal NVMe drive. Suspending goes OK, but the laptop doesn't resume properly. After much digging, it turns out that the _PS0 method for the NVMe drive is returning `1` - see line 7116 of https://bugzilla.kernel.org/attachment.cgi?id=228521 and the kernel informs that the device did not change to D0 from D3. I've been meaning to ask @l1k about this, but unfortunately I haven't had much spare time to work on this recently.

Of course, if you're running from an external USB drive, this doesn't really apply; but perhaps there's a simple cmdline flag I'm missing.