@cgutman 2016-09-12
I've been running from a USB disk, so I haven't hit the NVMe resume issue. I'm on 4.8-rc5 where I did hit a deadlock on suspend in brcmfmac which someone already found and fixed here: https://patchwork.kernel.org/patch/9280681/

I still do see the message complaining about a device not transitioning from D3 to D0 on resume, so I don't think it's been fixed in mainline yet. Since I wasn't using the NVMe drive, I never noticed it dropping off the bus after resume.