@l1k 2016-09-12
@cgutman: You're calling SIEN on `->probe`, not sure why this is needed, but I could imagine this makes sense e.g. if the driver was not loaded on boot (blacklisted), the machine suspended and the driver subsequently loaded manually.

However on the MBP12,1, if the controller was switched to USB mode by EFI and the USB driver has already loaded, things will break because you're taking the interface away from it. You need to test presence of UIST, if it exists call it and if the result is 1, bail out. And only *afterwards* call SIEN.

Please store the handle to the SIEN method in `struct applespi_data` so that the namespace doesn't have to be searched for it on every resume. (See l1k/linux@d65f187ebfc302dc516a08c14590ed578bce1b71 for an example.)

The 50 ms delay is quite long, is this really necessary? The macOS driver doesn't seem to delay at all AFAICS, but maybe I've missed something.

With that addressed,
`Reviewed-by: Lukas Wunner <lukas@wunner.de>`