@l1k 2016-09-12
@cb22: That's not the _PS0 method of the NVMe drive but of the root port it's attached to. It checks the Data Link Active bit in the PCIe Link Status register and bails out if the bit doesn't revert to 1. What does `lspci` say before and after resume for the root port and the NVMe drive below it?

`lspci -vvvv -s 00:1c.0 # RP01`
`lspci -vvvv -s 01:00.0 # SSD0, not sure if the bus number is correct`