@l1k 2016-09-12
The behaviour of _PS0 changes depending on a byte named NVME in the GNVS region. The byte should be at address 0x8AF9C918 + offset 0xB2 + 42 bytes = 0x8AF9C9F4 (decimal 2331625972). Which value do you get if you type:
`dd iflag=skip_bytes,count_bytes skip=2331625972 count=1 if=/dev/mem 2>/dev/null | hexdump`

As a shot in the dark, EFI configures the machine in a special BootCamp mode for Windows unless the OS identifies as macOS using a proprietary EFI protocol before ExitBootServices is called. Perhaps the SSD is treated differently on suspend/resume depending on macOS vs. BootCamp mode. You could try one of the methods described at https://github.com/0xbb/gpu-switch#macbook-pro-113-and-115-notes to enable macOS mode and see if this fixes suspend/resume for the SSD.